Move iceberg session properties to hive common module
