+++ presto-delta/src/main/java/com/facebook/presto/delta/DeltaConnectorFactory.java
@@ -16,6 +16,7 @@ package com.facebook.presto.delta;
 import com.facebook.airlift.bootstrap.Bootstrap;
 import com.facebook.airlift.json.JsonModule;
 import com.facebook.presto.cache.CachingModule;
+import com.facebook.presto.hive.HiveCommonModule;
 import com.facebook.presto.hive.authentication.HiveAuthenticationModule;
 import com.facebook.presto.hive.gcs.HiveGcsModule;
 import com.facebook.presto.hive.metastore.HiveMetastoreModule;
@@ -61,6 +62,7 @@ public class DeltaConnectorFactory
                     new HiveGcsModule(),
                     new HiveAuthenticationModule(),
                     new HiveMetastoreModule(catalogName, Optional.empty()),
+                    new HiveCommonModule(),
                     binder -> {
                         binder.bind(RowExpressionService.class).toInstance(context.getRowExpressionService());
                     });
