+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveClientConfig.java
@@ -19,8 +19,6 @@ import com.facebook.airlift.configuration.DefunctConfig;
 import com.facebook.airlift.configuration.LegacyConfig;
 import com.facebook.drift.transport.netty.codec.Protocol;
 import com.facebook.presto.hive.s3.S3FileSystemType;
-import com.facebook.presto.orc.OrcWriteValidation.OrcWriteValidationMode;
-import com.facebook.presto.spi.schedule.NodeSelectionStrategy;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 import io.airlift.units.DataSize;
@@ -46,7 +44,6 @@ import static com.facebook.presto.hive.HiveClientConfig.InsertExistingPartitions
 import static com.facebook.presto.hive.HiveClientConfig.InsertExistingPartitionsBehavior.OVERWRITE;
 import static com.facebook.presto.hive.HiveSessionProperties.INSERT_EXISTING_PARTITIONS_BEHAVIOR;
 import static com.facebook.presto.hive.HiveStorageFormat.ORC;
-import static com.facebook.presto.spi.schedule.NodeSelectionStrategy.NO_PREFERENCE;
 import static com.google.common.base.Preconditions.checkArgument;
 import static io.airlift.units.DataSize.Unit.BYTE;
 import static io.airlift.units.DataSize.Unit.MEGABYTE;
@@ -78,7 +75,6 @@ public class HiveClientConfig
     private DataSize maxInitialSplitSize;
     private int domainCompactionThreshold = 100;
     private DataSize writerSortBufferSize = new DataSize(64, MEGABYTE);
-    private NodeSelectionStrategy nodeSelectionStrategy = NO_PREFERENCE;
     private boolean recursiveDirWalkerEnabled;
 
     private int maxConcurrentFileRenames = 20;
@@ -110,28 +106,11 @@ public class HiveClientConfig
     private List<String> resourceConfigFiles = ImmutableList.of();
 
     private DataSize textMaxLineLength = new DataSize(100, MEGABYTE);
-
-    private boolean useParquetColumnNames;
-    private DataSize parquetMaxReadBlockSize = new DataSize(16, MEGABYTE);
-
     private boolean assumeCanonicalPartitionKeys;
-
     private boolean useOrcColumnNames;
-    private boolean orcBloomFiltersEnabled;
     private double orcDefaultBloomFilterFpp = 0.05;
-    private DataSize orcMaxMergeDistance = new DataSize(1, MEGABYTE);
-    private DataSize orcMaxBufferSize = new DataSize(8, MEGABYTE);
-    private DataSize orcTinyStripeThreshold = new DataSize(8, MEGABYTE);
-    private DataSize orcStreamBufferSize = new DataSize(8, MEGABYTE);
-    private DataSize orcMaxReadBlockSize = new DataSize(16, MEGABYTE);
-    private boolean orcLazyReadSmallRanges = true;
-    private boolean orcOptimizedWriterEnabled = true;
-    private double orcWriterValidationPercentage;
-    private OrcWriteValidationMode orcWriterValidationMode = OrcWriteValidationMode.BOTH;
-
     private boolean rcfileOptimizedWriterEnabled = true;
     private boolean rcfileWriterValidate;
-
     private HdfsAuthenticationType hdfsAuthenticationType = HdfsAuthenticationType.NONE;
     private boolean hdfsImpersonationEnabled;
     private boolean hdfsWireEncryptionEnabled;
@@ -178,15 +157,11 @@ public class HiveClientConfig
     private boolean pushdownFilterEnabled;
     private boolean parquetPushdownFilterEnabled;
     private boolean adaptiveFilterReorderingEnabled = true;
-    private boolean zstdJniDecompressionEnabled;
-
     private Duration fileStatusCacheExpireAfterWrite = new Duration(0, TimeUnit.SECONDS);
     private long fileStatusCacheMaxSize;
     private List<String> fileStatusCacheTables = ImmutableList.of();
 
     private DataSize pageFileStripeMaxSize = new DataSize(24, MEGABYTE);
-    private boolean parquetBatchReadOptimizationEnabled;
-    private boolean parquetEnableBatchReaderVerification;
     private boolean parquetDereferencePushdownEnabled;
 
     private int maxMetadataUpdaterThreads = 100;
@@ -219,7 +194,6 @@ public class HiveClientConfig
     private boolean fileSplittable = true;
     private Protocol thriftProtocol = Protocol.BINARY;
     private DataSize thriftBufferSize = new DataSize(128, BYTE);
-    private boolean isReadNullMaskedParquetEncryptedValueEnabled;
 
     private boolean copyOnFirstWriteConfigurationEnabled = true;
 
@@ -294,18 +268,6 @@ public class HiveClientConfig
     public HiveClientConfig setWriterSortBufferSize(DataSize writerSortBufferSize)
     {
         this.writerSortBufferSize = writerSortBufferSize;
-        return this;
-    }
-
-    public NodeSelectionStrategy getNodeSelectionStrategy()
-    {
-        return nodeSelectionStrategy;
-    }
-
-    @Config("hive.node-selection-strategy")
-    public HiveClientConfig setNodeSelectionStrategy(NodeSelectionStrategy nodeSelectionStrategy)
-    {
-        this.nodeSelectionStrategy = nodeSelectionStrategy;
         return this;
     }
 
@@ -779,99 +741,6 @@ public class HiveClientConfig
         return this;
     }
 
-    @NotNull
-    public DataSize getOrcMaxMergeDistance()
-    {
-        return orcMaxMergeDistance;
-    }
-
-    @Config("hive.orc.max-merge-distance")
-    public HiveClientConfig setOrcMaxMergeDistance(DataSize orcMaxMergeDistance)
-    {
-        this.orcMaxMergeDistance = orcMaxMergeDistance;
-        return this;
-    }
-
-    @NotNull
-    public DataSize getOrcMaxBufferSize()
-    {
-        return orcMaxBufferSize;
-    }
-
-    @Config("hive.orc.max-buffer-size")
-    public HiveClientConfig setOrcMaxBufferSize(DataSize orcMaxBufferSize)
-    {
-        this.orcMaxBufferSize = orcMaxBufferSize;
-        return this;
-    }
-
-    @NotNull
-    public DataSize getOrcStreamBufferSize()
-    {
-        return orcStreamBufferSize;
-    }
-
-    @Config("hive.orc.stream-buffer-size")
-    public HiveClientConfig setOrcStreamBufferSize(DataSize orcStreamBufferSize)
-    {
-        this.orcStreamBufferSize = orcStreamBufferSize;
-        return this;
-    }
-
-    @NotNull
-    public DataSize getOrcTinyStripeThreshold()
-    {
-        return orcTinyStripeThreshold;
-    }
-
-    @Config("hive.orc.tiny-stripe-threshold")
-    public HiveClientConfig setOrcTinyStripeThreshold(DataSize orcTinyStripeThreshold)
-    {
-        this.orcTinyStripeThreshold = orcTinyStripeThreshold;
-        return this;
-    }
-
-    @NotNull
-    public DataSize getOrcMaxReadBlockSize()
-    {
-        return orcMaxReadBlockSize;
-    }
-
-    @Config("hive.orc.max-read-block-size")
-    public HiveClientConfig setOrcMaxReadBlockSize(DataSize orcMaxReadBlockSize)
-    {
-        this.orcMaxReadBlockSize = orcMaxReadBlockSize;
-        return this;
-    }
-
-    @Deprecated
-    public boolean isOrcLazyReadSmallRanges()
-    {
-        return orcLazyReadSmallRanges;
-    }
-
-    // TODO remove config option once efficacy is proven
-    @Deprecated
-    @Config("hive.orc.lazy-read-small-ranges")
-    @ConfigDescription("ORC read small disk ranges lazily")
-    public HiveClientConfig setOrcLazyReadSmallRanges(boolean orcLazyReadSmallRanges)
-    {
-        this.orcLazyReadSmallRanges = orcLazyReadSmallRanges;
-        return this;
-    }
-
-    public boolean isOrcBloomFiltersEnabled()
-    {
-        return orcBloomFiltersEnabled;
-    }
-
-    @Config("hive.orc.bloom-filters.enabled")
-    public HiveClientConfig setOrcBloomFiltersEnabled(boolean orcBloomFiltersEnabled)
-    {
-        this.orcBloomFiltersEnabled = orcBloomFiltersEnabled;
-        return this;
-    }
-
     public double getOrcDefaultBloomFilterFpp()
     {
         return orcDefaultBloomFilterFpp;
@@ -882,49 +751,6 @@ public class HiveClientConfig
     public HiveClientConfig setOrcDefaultBloomFilterFpp(double orcDefaultBloomFilterFpp)
     {
         this.orcDefaultBloomFilterFpp = orcDefaultBloomFilterFpp;
-        return this;
-    }
-
-    @Deprecated
-    public boolean isOrcOptimizedWriterEnabled()
-    {
-        return orcOptimizedWriterEnabled;
-    }
-
-    @Deprecated
-    @Config("hive.orc.optimized-writer.enabled")
-    public HiveClientConfig setOrcOptimizedWriterEnabled(boolean orcOptimizedWriterEnabled)
-    {
-        this.orcOptimizedWriterEnabled = orcOptimizedWriterEnabled;
-        return this;
-    }
-
-    @DecimalMin("0.0")
-    @DecimalMax("100.0")
-    public double getOrcWriterValidationPercentage()
-    {
-        return orcWriterValidationPercentage;
-    }
-
-    @Config("hive.orc.writer.validation-percentage")
-    @ConfigDescription("Percentage of ORC files to validate after write by re-reading the whole file")
-    public HiveClientConfig setOrcWriterValidationPercentage(double orcWriterValidationPercentage)
-    {
-        this.orcWriterValidationPercentage = orcWriterValidationPercentage;
-        return this;
-    }
-
-    @NotNull
-    public OrcWriteValidationMode getOrcWriterValidationMode()
-    {
-        return orcWriterValidationMode;
-    }
-
-    @Config("hive.orc.writer.validation-mode")
-    @ConfigDescription("Level of detail in ORC validation. Lower levels require more memory.")
-    public HiveClientConfig setOrcWriterValidationMode(OrcWriteValidationMode orcWriterValidationMode)
-    {
-        this.orcWriterValidationMode = orcWriterValidationMode;
         return this;
     }
 
@@ -983,32 +809,6 @@ public class HiveClientConfig
         return this;
     }
 
-    public boolean isUseParquetColumnNames()
-    {
-        return useParquetColumnNames;
-    }
-
-    @Config("hive.parquet.use-column-names")
-    @ConfigDescription("Access Parquet columns using names from the file")
-    public HiveClientConfig setUseParquetColumnNames(boolean useParquetColumnNames)
-    {
-        this.useParquetColumnNames = useParquetColumnNames;
-        return this;
-    }
-
-    @NotNull
-    public DataSize getParquetMaxReadBlockSize()
-    {
-        return parquetMaxReadBlockSize;
-    }
-
-    @Config("hive.parquet.max-read-block-size")
-    public HiveClientConfig setParquetMaxReadBlockSize(DataSize parquetMaxReadBlockSize)
-    {
-        this.parquetMaxReadBlockSize = parquetMaxReadBlockSize;
-        return this;
-    }
-
     @Deprecated
     public boolean isOptimizeMismatchedBucketCount()
     {
@@ -1020,18 +820,6 @@ public class HiveClientConfig
     public HiveClientConfig setOptimizeMismatchedBucketCount(boolean optimizeMismatchedBucketCount)
     {
         this.optimizeMismatchedBucketCount = optimizeMismatchedBucketCount;
-        return this;
-    }
-
-    public boolean isZstdJniDecompressionEnabled()
-    {
-        return zstdJniDecompressionEnabled;
-    }
-
-    @Config("hive.zstd-jni-decompression-enabled")
-    public HiveClientConfig setZstdJniDecompressionEnabled(boolean zstdJniDecompressionEnabled)
-    {
-        this.zstdJniDecompressionEnabled = zstdJniDecompressionEnabled;
         return this;
     }
 
@@ -1550,32 +1338,6 @@ public class HiveClientConfig
         return this;
     }
 
-    @Config("hive.parquet-batch-read-optimization-enabled")
-    @ConfigDescription("enable parquet batch reads optimization")
-    public HiveClientConfig setParquetBatchReadOptimizationEnabled(boolean parquetBatchReadOptimizationEnabled)
-    {
-        this.parquetBatchReadOptimizationEnabled = parquetBatchReadOptimizationEnabled;
-        return this;
-    }
-
-    public boolean isParquetBatchReadOptimizationEnabled()
-    {
-        return this.parquetBatchReadOptimizationEnabled;
-    }
-
-    @Config("hive.enable-parquet-batch-reader-verification")
-    @ConfigDescription("enable optimized parquet reader")
-    public HiveClientConfig setParquetBatchReaderVerificationEnabled(boolean parquetEnableBatchReaderVerification)
-    {
-        this.parquetEnableBatchReaderVerification = parquetEnableBatchReaderVerification;
-        return this;
-    }
-
-    public boolean isParquetBatchReaderVerificationEnabled()
-    {
-        return this.parquetEnableBatchReaderVerification;
-    }
-
     @Config("hive.enable-parquet-dereference-pushdown")
     @ConfigDescription("enable parquet dereference pushdown")
     public HiveClientConfig setParquetDereferencePushdownEnabled(boolean parquetDereferencePushdownEnabled)
@@ -1874,19 +1636,6 @@ public class HiveClientConfig
     {
         this.partitionFilteringFromMetastoreEnabled = partitionFilteringFromMetastoreEnabled;
         return this;
-    }
-
-    @Config("hive.read-null-masked-parquet-encrypted-value-enabled")
-    @ConfigDescription("Read null masked value when access is denied for an encrypted parquet column")
-    public HiveClientConfig setReadNullMaskedParquetEncryptedValue(boolean isReadNullMaskedParquetEncryptedValueEnabled)
-    {
-        this.isReadNullMaskedParquetEncryptedValueEnabled = isReadNullMaskedParquetEncryptedValueEnabled;
-        return this;
-    }
-
-    public boolean getReadNullMaskedParquetEncryptedValue()
-    {
-        return this.isReadNullMaskedParquetEncryptedValueEnabled;
     }
 
     @Config("hive.parallel-parsing-of-partition-values-enabled")
