+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveClientModule.java
@@ -123,10 +123,8 @@ public class HiveClientModule
         binder.bind(HdfsConfigurationInitializer.class).in(Scopes.SINGLETON);
         newSetBinder(binder, DynamicConfigurationProvider.class);
         configBinder(binder).bindConfig(HiveClientConfig.class);
-        configBinder(binder).bindConfig(HiveCommonClientConfig.class);
 
         binder.bind(HiveSessionProperties.class).in(Scopes.SINGLETON);
-        binder.bind(HiveCommonSessionProperties.class).in(Scopes.SINGLETON);
         binder.bind(HiveTableProperties.class).in(Scopes.SINGLETON);
         binder.bind(HiveAnalyzeProperties.class).in(Scopes.SINGLETON);
