+++ presto-hive/src/test/java/com/facebook/presto/hive/TestHiveSessionProperties.java
@@ -19,7 +19,7 @@ import com.facebook.presto.spi.schedule.NodeSelectionStrategy;
 import com.facebook.presto.testing.TestingConnectorSession;
 import org.testng.annotations.Test;
 
-import static com.facebook.presto.hive.HiveSessionProperties.getNodeSelectionStrategy;
+import static com.facebook.presto.hive.HiveCommonSessionProperties.getNodeSelectionStrategy;
 import static com.facebook.presto.hive.HiveSessionProperties.getParquetWriterVersion;
 import static com.facebook.presto.hive.HiveSessionProperties.isCacheEnabled;
 import static com.facebook.presto.spi.schedule.NodeSelectionStrategy.HARD_AFFINITY;
@@ -35,11 +35,8 @@ public class TestHiveSessionProperties
     public void testEmptyNodeSelectionStrategyConfig()
     {
         ConnectorSession connectorSession = new TestingConnectorSession(
-                new HiveSessionProperties(
-                        new HiveClientConfig(),
-                        new OrcFileWriterConfig(),
-                        new ParquetFileWriterConfig(),
-                        new CacheConfig()).getSessionProperties());
+                new HiveCommonSessionProperties(
+                        new HiveCommonClientConfig()).getSessionProperties());
         assertEquals(getNodeSelectionStrategy(connectorSession), NO_PREFERENCE);
     }
 
@@ -47,11 +44,9 @@ public class TestHiveSessionProperties
     public void testEmptyConfigNodeSelectionStrategyConfig()
     {
         ConnectorSession connectorSession = new TestingConnectorSession(
-                new HiveSessionProperties(
-                        new HiveClientConfig().setNodeSelectionStrategy(NodeSelectionStrategy.valueOf("NO_PREFERENCE")),
-                        new OrcFileWriterConfig(),
-                        new ParquetFileWriterConfig(),
-                        new CacheConfig()).getSessionProperties());
+                new HiveCommonSessionProperties(
+                        new HiveCommonClientConfig().setNodeSelectionStrategy(NodeSelectionStrategy.valueOf("NO_PREFERENCE"))
+                ).getSessionProperties());
         assertEquals(getNodeSelectionStrategy(connectorSession), NO_PREFERENCE);
     }
 
@@ -59,11 +54,9 @@ public class TestHiveSessionProperties
     public void testNodeSelectionStrategyConfig()
     {
         ConnectorSession connectorSession = new TestingConnectorSession(
-                new HiveSessionProperties(
-                        new HiveClientConfig().setNodeSelectionStrategy(HARD_AFFINITY),
-                        new OrcFileWriterConfig(),
-                        new ParquetFileWriterConfig(),
-                        new CacheConfig()).getSessionProperties());
+                new HiveCommonSessionProperties(
+                        new HiveCommonClientConfig().setNodeSelectionStrategy(HARD_AFFINITY)
+                ).getSessionProperties());
         assertEquals(getNodeSelectionStrategy(connectorSession), HARD_AFFINITY);
     }
