+++ presto-hudi/src/main/java/com/facebook/presto/hudi/HudiSessionProperties.java
@@ -14,7 +14,6 @@
 
 package com.facebook.presto.hudi;
 
-import com.facebook.presto.hive.HiveClientConfig;
 import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.session.PropertyMetadata;
@@ -27,7 +26,6 @@ import java.util.List;
 
 import static com.facebook.presto.common.type.DoubleType.DOUBLE;
 import static com.facebook.presto.hive.HiveSessionProperties.CACHE_ENABLED;
-import static com.facebook.presto.hive.HiveSessionProperties.dataSizeSessionProperty;
 import static com.facebook.presto.spi.StandardErrorCode.INVALID_SESSION_PROPERTY;
 import static com.facebook.presto.spi.session.PropertyMetadata.booleanProperty;
 import static com.facebook.presto.spi.session.PropertyMetadata.dataSizeProperty;
@@ -39,20 +37,14 @@ public class HudiSessionProperties
     private final List<PropertyMetadata<?>> sessionProperties;
 
     private static final String HUDI_METADATA_TABLE_ENABLED = "hudi_metadata_table_enabled";
-
-    private static final String PARQUET_MAX_READ_BLOCK_SIZE = "parquet_max_read_block_size";
-    private static final String PARQUET_BATCH_READ_OPTIMIZATION_ENABLED = "parquet_batch_read_optimization_enabled";
-    private static final String PARQUET_BATCH_READER_VERIFICATION_ENABLED = "parquet_batch_reader_verification_enabled";
-
     private static final String SIZE_BASED_SPLIT_WEIGHTS_ENABLED = "size_based_split_weights_enabled";
     private static final String STANDARD_SPLIT_WEIGHT_SIZE = "standard_split_weight_size";
     private static final String MINIMUM_ASSIGNED_SPLIT_WEIGHT = "minimum_assigned_split_weight";
-    public static final String READ_MASKED_VALUE_ENABLED = "read_null_masked_parquet_encrypted_value_enabled";
     private static final String MAX_OUTSTANDING_SPLITS = "max_outstanding_splits";
     private static final String SPLIT_GENERATOR_PARALLELISM = "split_generator_parallelism";
 
     @Inject
-    public HudiSessionProperties(HiveClientConfig hiveClientConfig, HudiConfig hudiConfig)
+    public HudiSessionProperties(HudiConfig hudiConfig)
     {
         sessionProperties = ImmutableList.of(
                 booleanProperty(
@@ -66,21 +58,6 @@ public class HudiSessionProperties
                         "Enable cache for hive",
                         false,
                         true),
-                dataSizeSessionProperty(
-                        PARQUET_MAX_READ_BLOCK_SIZE,
-                        "Parquet: Maximum size of a block to read",
-                        hiveClientConfig.getParquetMaxReadBlockSize(),
-                        false),
-                booleanProperty(
-                        PARQUET_BATCH_READ_OPTIMIZATION_ENABLED,
-                        "Is Parquet batch read optimization enabled",
-                        false,
-                        false),
-                booleanProperty(
-                        PARQUET_BATCH_READER_VERIFICATION_ENABLED,
-                        "Is Parquet batch reader verification enabled? This is for testing purposes only, not to be used in production",
-                        false,
-                        false),
                 booleanProperty(
                         SIZE_BASED_SPLIT_WEIGHTS_ENABLED,
                         format("If enabled, size-based splitting ensures that each batch of splits has enough data to process as defined by %s", STANDARD_SPLIT_WEIGHT_SIZE),
@@ -106,11 +83,6 @@ public class HudiSessionProperties
                             return doubleValue;
                         },
                         value -> value),
-                booleanProperty(
-                        READ_MASKED_VALUE_ENABLED,
-                        "Return null when access is denied for an encrypted parquet column",
-                        hiveClientConfig.getReadNullMaskedParquetEncryptedValue(),
-                        false),
                 integerProperty(
                         MAX_OUTSTANDING_SPLITS,
                         "Maximum outstanding splits per batch for query",
@@ -133,21 +105,6 @@ public class HudiSessionProperties
         return session.getProperty(HUDI_METADATA_TABLE_ENABLED, Boolean.class);
     }
 
-    public static DataSize getParquetMaxReadBlockSize(ConnectorSession session)
-    {
-        return session.getProperty(PARQUET_MAX_READ_BLOCK_SIZE, DataSize.class);
-    }
-
-    public static boolean isParquetBatchReadsEnabled(ConnectorSession session)
-    {
-        return session.getProperty(PARQUET_BATCH_READ_OPTIMIZATION_ENABLED, Boolean.class);
-    }
-
-    public static boolean isParquetBatchReaderVerificationEnabled(ConnectorSession session)
-    {
-        return session.getProperty(PARQUET_BATCH_READER_VERIFICATION_ENABLED, Boolean.class);
-    }
-
     public static boolean isSizeBasedSplitWeightsEnabled(ConnectorSession session)
     {
         return session.getProperty(SIZE_BASED_SPLIT_WEIGHTS_ENABLED, Boolean.class);
@@ -161,11 +118,6 @@ public class HudiSessionProperties
     public static double getMinimumAssignedSplitWeight(ConnectorSession session)
     {
         return session.getProperty(MINIMUM_ASSIGNED_SPLIT_WEIGHT, Double.class);
-    }
-
-    public static boolean getReadNullMaskedParquetEncryptedValue(ConnectorSession session)
-    {
-        return session.getProperty(READ_MASKED_VALUE_ENABLED, Boolean.class);
     }
 
     public static int getMaxOutstandingSplits(ConnectorSession session)
